refactor: delete unused filse and apply lint
