+++ front/lib/core/project/data/models/project.dart
@@ -1,3 +1,5 @@
+// ignore_for_file: non_constant_identifier_names
+
 import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:front/core/project/domain/entities/project.dart';
