+++ front/lib/core/project/domain/entities/project.dart
@@ -1,3 +1,5 @@
+// ignore_for_file: non_constant_identifier_names
+
 import 'package:equatable/equatable.dart';
 
 class Project extends Equatable {
