+++ front/lib/shared/atom/timer.dart
@@ -14,7 +14,7 @@ class TimerWidget extends StatefulWidget {
 
 
   @override
-  _TimerWidgetState createState() => _TimerWidgetState();
+  State<TimerWidget> createState() => _TimerWidgetState();
 }
 
 class _TimerWidgetState extends State<TimerWidget> {
