+++ front/lib/shared/helper/FormHelper/form_helper.dart
@@ -1,3 +1,2 @@
 export './form_validate_builder.dart';
 export 'interface/form_error_type.dart';
-export 'interface/form_validate_returnType.dart';
