--- front/lib/shared/helper/FormHelper/interface/form_validate_returnType.dart
@@ -1,5 +1,0 @@
-enum ValidateReturnType {
-  onChange,
-  onValidate,
-}
-typedef ValidateFunction = String Function(String?);
